payroll_analyst/requirements.txt
requests==2.22.0
numpy==1.17.3
python-dotenv==0.10.3
terminaltables==3.1.0